file_sync/requirements.txt
psycopg2-binary
requests
websocket-client
flask
